Refactoring
relocate gift data storage out of legal.cs
