+++ PKHeX.Core/Legality/Encounters/Generator/EncounterGenerator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using static PKHeX.Core.Legal;
+using static PKHeX.Core.EncounterEvent;
 
 namespace PKHeX.Core
 {
