+++ PKHeX.Core/Legality/Learnset/Learnset6.cs
@@ -1,0 +1,29 @@
+using System.IO;
+
+namespace PKHeX.Core
+{
+    public class Learnset6 : Learnset
+    {
+        private Learnset6(byte[] data)
+        {
+            if (data.Length < 4 || data.Length % 4 != 0)
+            { Count = 0; Levels = new int[0]; Moves = new int[0]; return; }
+            Count = data.Length / 4 - 1;
+            Moves = new int[Count];
+            Levels = new int[Count];
+            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
+                for (int i = 0; i < Count; i++)
+                {
+                    Moves[i] = br.ReadInt16();
+                    Levels[i] = br.ReadInt16();
+                }
+        }
+        public static Learnset[] GetArray(byte[][] entries)
+        {
+            Learnset[] data = new Learnset[entries.Length];
+            for (int i = 0; i < data.Length; i++)
+                data[i] = new Learnset6(entries[i]);
+            return data;
+        }
+    }
+}
