+++ PKHeX.WinForms/Subforms/SAV_MysteryGiftDB.cs
@@ -203,14 +203,7 @@ namespace PKHeX.WinForms
         }
         private void LoadDatabase()
         {
-            RawDB = new List<MysteryGift>();
-            RawDB.AddRange(Legal.MGDB_G4);
-            RawDB.AddRange(Legal.MGDB_G5);
-            RawDB.AddRange(Legal.MGDB_G6);
-            RawDB.AddRange(Legal.MGDB_G7);
-
-            RawDB = new List<MysteryGift>(RawDB.Where(mg => !mg.IsItem && mg.IsPokémon && mg.Species > 0).Distinct()
-                .Concat(Legal.MGDB_G3).OrderBy(mg => mg.Species));
+            RawDB = new List<MysteryGift>(EncounterEvent.GetAllEvents());
             foreach (var mg in RawDB)
                 mg.GiftUsed = false;
             BeginInvoke(new MethodInvoker(delegate
